+++ src/test/resources/inspections/ForeachToCollectionInspection/foreach-outer-scope-variable-used-twice.after.php
@@ -1,0 +1,15 @@
+<?php
+
+$array = [
+    'one',
+    'two',
+    'tree'
+];
+
+$name = 'one';
+
+collect($array)->each(function ($item) use ($name) {
+    if ($item === $name) {
+        echo "found " . $name;
+    }
+});