+++ CHANGELOG.md
@@ -12,8 +12,10 @@
 ### Removed
 
 ### Fixed
+- Fixed foreach refactoring. `use` statement missing array variable when used in foreach.
+- Fixed foreach refactoring. `use` statement having duplicate variable, if used twice.
+- Fixed foreach refactoring. `use` statement containing array variable when using property accessor.
 
-### Security
 ## [0.0.1-EAP.5]
 ### Fixed
 - Fixed refactoring foreach to collection with multiline